requirements.txt
gitpython>=3.1.32
Markdown==3.4.4
ollama>=0.4.1
pelican==4.8.0
